pyproject.toml
[project]
name = "finz-analysis-model"
version = "0.1.0"
description = "Multimodal deep learning model for stock momentum prediction from news text and financial data"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.35",
    "tokenizers>=0.15",
    "datasets>=2.14",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "yfinance>=0.2.30",
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "tqdm>=4.65",
    "wandb>=0.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "ipykernel>=6.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
